requirements.txt
openai
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation
opentelemetry-exporter-otlp
python-dotenv
opentelemetry-instrumentation-requests